Adds activity evaluation popup to profile page
Implements a popup that allows users to evaluate participants
of past activities.

The popup includes a list of participants and a rating component,
allowing the user to assign a rating to each participant.

The rating is temporarily stored in the state to show which rating
is given for each participant.

Adds basic styling for the popup and its elements.